requirements.txt
matplotlib~=3.3.4
numpy~=1.20.2
Pillow~=8.2.0
PyYAML~=5.4.1
tensorflow==2.7.0
auto_mix_prep==0.2.0
yaml~=0.2.5
opencv-python~=4.4.0.46